Exception handling

+++ dscommerce/src/main/java/com/github/luizns/dscommerce/service/ProductService.java
@@ -3,10 +3,14 @@ package com.github.luizns.dscommerce.service;
 import com.github.luizns.dscommerce.DTO.ProductDTO;
 import com.github.luizns.dscommerce.entities.Product;
 import com.github.luizns.dscommerce.repository.ProductRepository;
+import com.github.luizns.dscommerce.service.exceptions.DatabaseException;
+import com.github.luizns.dscommerce.service.exceptions.ResourceNotFoundException;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.dao.DataIntegrityViolationException;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
 import org.springframework.stereotype.Service;
+import org.springframework.transaction.annotation.Propagation;
 import org.springframework.transaction.annotation.Transactional;
 
 @Service
@@ -17,7 +21,7 @@ public class ProductService {
 
     @Transactional(readOnly = true)
     public ProductDTO findById(Long id) {
-        Product product = repository.findById(id).get();
+        Product product = repository.findById(id).orElseThrow(() -> new ResourceNotFoundException("Recurso não encontrado"));
         return new ProductDTO(product);
     }
 
@@ -37,16 +41,28 @@ public class ProductService {
 
     @Transactional
     public ProductDTO update(Long id, ProductDTO dto) {
-        Product entity = repository.getReferenceById(id);
-        copyDtoToEntity(entity, dto);
-        entity = repository.save(entity);
-        return new ProductDTO(entity);
+        try {
+            Product entity = repository.getReferenceById(id);
+            copyDtoToEntity(entity, dto);
+            entity = repository.save(entity);
+            return new ProductDTO(entity);
+        } catch (Exception e) {
+            throw new ResourceNotFoundException("Recurso não encontrado");
+        }
     }
 
-    @Transactional
+    @Transactional(propagation = Propagation.SUPPORTS)
     public void delete(Long id) {
-        repository.deleteById(id);
+        if (!repository.existsById(id)) {
+            throw new ResourceNotFoundException("Recurso não encontrado");
+        }
+        try {
+            repository.deleteById(id);
+        } catch (DataIntegrityViolationException e) {
+            throw new DatabaseException("Falha de integridade referencial");
+        }
     }
+
 
     private void copyDtoToEntity(Product entity, ProductDTO dto) {
         entity.setName(dto.getName());
